feat: fractional grid_deg + grid combo box + None never competes
Container header grid_deg widens u8 -> f32 (decoder offsets shift by 3)
so fractional prefilter cells (0.5, 4/3, ...) ride through the whole
stack: Params/wasm export/CLI args become f64, encode validates 0.1-45.
Verified byte-identical wasm-vs-native at 2 and 0.5 degrees.

The encode panel's per-layer grid becomes a combo box (datalist with
the coarse pre-filter's detents + free-text 0.1-45), and the encode
worker gets an onerror handler so an OOM on a very fine grid surfaces
as an error row instead of silence.

'No zone' never competes in the coarse views: pre-filter dominance,
share, misassigned-km2 and the hover candidate mix now count only
zone-owned lattice samples, so a part-None cell reads 100% its zone --
matching the runtime, whose dominance tallies already exclude unowned
subcells (lookup_coarse only answers None for cells with no zone at
all).

+++ utz-build/src/cmd/encode.rs
@@ -26,8 +26,8 @@ pub struct Args {
     #[arg(long, default_value_t = 24)]
     qbits: u32,
     /// grid cell size in integer degrees
-    #[arg(long, default_value_t = 2)]
-    grid_deg: u32,
+    #[arg(long, default_value_t = 2.0)]
+    grid_deg: f64,
     /// enable population weighting with this floor multiplier (e.g. 0.052)
     #[arg(long)]
     w_min: Option<f64>,
